For EXP-2012: Moving tests for versioning into their own file
