+++ components/nimbus/src/tests/mod.rs
@@ -9,3 +9,4 @@ mod test_lib;
 mod test_lib_bw_compat;
 mod test_lib_schema_deserialization;
 mod test_persistence;
+mod test_versioning;
